ARROW-9355 [R]: Fix -Wimplicit-int-float-conversion
I wasn't able to reproduce the problem locally, but this should fix it

Closes #7703 from romainfrancois/ARROW-9355/cast

Authored-by: Romain Francois <romain@rstudio.com>

+++ r/src/array_from_vector.cpp
@@ -403,7 +403,8 @@ namespace internal {
 template <typename T, typename Target,
           typename std::enable_if<std::is_signed<Target>::value, Target>::type = 0>
 Status int_cast(T x, Target* out) {
-  if (x < std::numeric_limits<Target>::min() || x > std::numeric_limits<Target>::max()) {
+  if (static_cast<int64_t>(x) < std::numeric_limits<Target>::min() ||
+      static_cast<int64_t>(x) > std::numeric_limits<Target>::max()) {
     return Status::Invalid("Value is too large to fit in C integer type");
   }
   *out = static_cast<Target>(x);
